Materialize stub TaskFlow arg bindings in core Dag serialization
Review feedback flagged two issues with keeping this in the provider:
the execution API gated on @task.stub's operator class name, which any
future lang-SDK operator would have had to duplicate, and generating
JSON-schema fragments from Python type hints is tightly coupled to the
execution API and didn't belong in the standard provider's decorator.

A generic inherits_from_stub_operator flag (mirroring EmptyOperator's
inherits_from_empty_operator) now drives arg-binding materialization
from OperatorSerialization._serialize_node, so the provider's
_StubOperator shrinks to just its own structural checks and any future
foreign-runtime operator gets the same treatment for free.
