+++ airflow-core/src/airflow/serialization/definitions/baseoperator.py
@@ -68,6 +68,8 @@ class SerializedBaseOperator(DAGNode):
 
     _can_skip_downstream: bool
     _is_empty: bool
+    _is_stub_operator: bool = False
+    _arg_bindings: list[dict[str, Any]] | None = None
     _needs_expansion: bool
     _task_display_name: str | None = None
     _weight_rule: str | PriorityWeightStrategy = "downstream"
@@ -294,6 +296,10 @@ class SerializedBaseOperator(DAGNode):
     @property
     def inherits_from_skipmixin(self) -> bool:
         return self._can_skip_downstream
+
+    @property
+    def inherits_from_stub_operator(self) -> bool:
+        return self._is_stub_operator
 
     @property
     def operator_name(self) -> str:
